apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1
redis==5.2.1
minio==7.2.13
python-multipart==0.0.20
argon2-cffi==23.1.0
itsdangerous==2.2.0

# --- dev / test ---
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
